move remaining exceptions
