+++ py_jama_rest_client/client.py
@@ -1,55 +1,19 @@
 import json
 import logging
-
-from .core import Core, CoreException
+from py_jama_rest_client.exceptions import (
+    APIException,
+    CoreException,
+    ResourceNotFoundException,
+    AlreadyExistsException,
+    TooManyRequestsException,
+    APIClientException,
+    APIServerException,
+    UnauthorizedException,
+)
+from py_jama_rest_client.core import Core
 
 # This is the py_jama_rest_client logger.
 py_jama_rest_client_logger = logging.getLogger("py_jama_rest_client")
-
-
-class APIException(Exception):
-    """This is the base class for all exceptions raised by the JamaClient"""
-
-    def __init__(self, message, status_code=None, reason=None):
-        super(APIException, self).__init__(message)
-        self.status_code = status_code
-        self.reason = reason
-
-
-class UnauthorizedException(APIException):
-    """This exception is thrown whenever the api returns a 401 unauthorized response."""
-
-    pass
-
-
-class TooManyRequestsException(APIException):
-    """This exception is thrown whenever the api returns a 429 too many requests response."""
-
-    pass
-
-
-class ResourceNotFoundException(APIException):
-    """This exception is raised whenever the api returns a 404 not found response."""
-
-    pass
-
-
-class AlreadyExistsException(APIException):
-    """This exception is thrown when the API returns a 400 response with a message that the resource already exists."""
-
-    pass
-
-
-class APIClientException(APIException):
-    """This exception is thrown whenever a unknown 400 error is encountered."""
-
-    pass
-
-
-class APIServerException(APIException):
-    """This exception is thrown whenever an unknown 500 response is encountered."""
-
-    pass
 
 
 class JamaClient:
@@ -1615,7 +1579,6 @@ class JamaClient:
         except CoreException as err:
             py_jama_rest_client_logger.error(err)
             raise APIException(str(err))
-            raise APIException
         return self.__handle_response_status(response)
 
     def put_user_active(self, user_id, is_active):
